Move Item and Conversation from Core to Object
- Move BaseObject from Core\ to Friendica\
